requirements.txt
streamlit
python-dotenv
langchain-core
langchain-openai
huggingface-hub
requests
Pillow